\name{enrichvs-package}
\alias{enrichvs-package}
\alias{enrichvs}
\docType{package}
\title{
	Enrichment assessment of virtual screening approaches
}
\description{
	These programs are used for calculating enrichment factors,
	drawing enrichment curves to evaluate virtual screening approaches.
}
\details{
\tabular{ll}{
Package: \tab enrichvs\cr
Type: \tab Package\cr
Version: \tab 0.0.4\cr
Date: \tab 2011-05-03\cr
License: \tab BSD\cr
LazyLoad: \tab yes\cr
}
	In drug discovery processes, efficient virtual screening (VS) 
	is essential to select candidate compounds.
	This package provides an easy-to-use interface to evaluate VS approaches.
	Especially, plotting a enrichment curve helps us to find the best approach.
}
\author{
	Hiroaki YABUUCHI
}
\references{
	Yabuuchi H. et al. Analysis of multiple compound-protein interactions reveals novel bioactive molecules. Mol. Syst. Biol. (2011) 7, 472.
}

\keyword{ package }
\seealso{

}
\examples{

}
